Added separate file for store property
